@@ -7,17 +7,17 @@
     "start": "ts-node src/main"
   },
   "dependencies": {
-    "@nestjs/common": "6.0.3",
-    "@nestjs/core": "6.0.3",
-    "class-transformer": "0.2.0",
+    "@nestjs/common": "6.2.4",
+    "@nestjs/core": "6.2.4",
+    "class-transformer": "0.2.3",
     "class-validator": "0.9.1",
     "reflect-metadata": "0.1.13",
-    "rxjs": "6.4.0",
-    "typescript": "3.3.4000"
+    "rxjs": "6.5.2",
+    "typescript": "3.4.5"
   },
   "devDependencies": {
-    "@types/node": "7.10.5",
+    "@types/node": "7.10.6",
     "supertest": "4.0.2",
-    "ts-node": "8.0.3"
+    "ts-node": "8.1.0"
   }
 }
