@@ -9,18 +9,18 @@
     "start:prod": "node dist/main.js"
   },
   "dependencies": {
-    "@nestjs/common": "6.0.3",
-    "@nestjs/core": "6.0.3",
-    "@nestjs/mongoose": "6.0.0",
-    "mongoose": "5.4.20",
+    "@nestjs/common": "6.2.4",
+    "@nestjs/core": "6.2.4",
+    "@nestjs/mongoose": "6.1.2",
+    "mongoose": "5.5.10",
     "reflect-metadata": "0.1.13",
-    "rxjs": "6.4.0",
-    "typescript": "3.3.4000"
+    "rxjs": "6.5.2",
+    "typescript": "3.4.5"
   },
   "devDependencies": {
-    "@types/mongoose": "5.3.24",
-    "@types/node": "7.10.5",
-    "ts-node": "8.0.3",
-    "tslint": "5.14.0"
+    "@types/mongoose": "5.5.1",
+    "@types/node": "7.10.6",
+    "ts-node": "8.1.0",
+    "tslint": "5.16.0"
   }
 }
