@@ -9,18 +9,18 @@
     "start:prod": "node dist/main.js"
   },
   "dependencies": {
-    "@nestjs/common": "6.0.3",
-    "@nestjs/core": "6.0.3",
-    "@nestjs/typeorm": "6.0.0",
-    "mysql": "2.16.0",
+    "@nestjs/common": "6.2.4",
+    "@nestjs/core": "6.2.4",
+    "@nestjs/typeorm": "6.1.1",
+    "mysql": "2.17.1",
     "reflect-metadata": "0.1.13",
-    "rxjs": "6.4.0",
-    "typeorm": "0.2.16",
-    "typescript": "3.3.4000"
+    "rxjs": "6.5.2",
+    "typeorm": "0.2.17",
+    "typescript": "3.4.5"
   },
   "devDependencies": {
-    "@types/node": "7.10.5",
-    "ts-node": "8.0.3",
-    "tslint": "5.14.0"
+    "@types/node": "7.10.6",
+    "ts-node": "8.1.0",
+    "tslint": "5.16.0"
   }
 }
