@@ -1,6 +1,6 @@
-import { HelloService } from './hello.service';
 import { Controller, Get, Header, Param } from '@nestjs/common';
 import { Observable, of } from 'rxjs';
+import { HelloService } from './hello.service';
 import { UserByIdPipe } from './users/user-by-id.pipe';
 
 @Controller('hello')
