@@ -1,2 +1,2 @@
-export * from './validation.pipe';
 export * from './parse-int.pipe';
+export * from './validation.pipe';
