@@ -80,7 +80,7 @@ export class Injector {
     }
     const loadInstance = (instances: any[]) => {
       targetWrapper.instance = targetWrapper.isDependencyTreeStatic()
-        ? new metatype(...instances)
+        ? new (metatype as Type<any>)(...instances)
         : Object.create(metatype.prototype);
     };
     await this.resolveConstructorParams(
@@ -190,7 +190,7 @@ export class Injector {
       throw new RuntimeException();
     }
     if (instanceHost.isResolved) {
-      return;
+      return done();
     }
     const callback = async (instances: any[]) => {
       const properties = await this.resolveProperties(
@@ -235,10 +235,10 @@ export class Injector {
       return callback(deps);
     }
     const dependencies = isNil(inject)
-      ? this.reflectConstructorParams(wrapper.metatype)
+      ? this.reflectConstructorParams(wrapper.metatype as Type<any>)
       : inject;
     const optionalDependenciesIds = isNil(inject)
-      ? this.reflectOptionalParams(wrapper.metatype)
+      ? this.reflectOptionalParams(wrapper.metatype as Type<any>)
       : [];
 
     let isResolved = true;
@@ -504,7 +504,7 @@ export class Injector {
     if (metadata && contextId !== STATIC_CONTEXT) {
       return this.loadPropertiesMetadata(metadata, contextId, inquirer);
     }
-    const properties = this.reflectProperties(wrapper.metatype);
+    const properties = this.reflectProperties(wrapper.metatype as Type<any>);
     const instances = await Promise.all(
       properties.map(async (item: PropertyDependency) => {
         try {
@@ -588,14 +588,12 @@ export class Injector {
     const isInContext = isStatic || isInRequestScope || isLazyTransient;
 
     if (isNil(inject) && isInContext) {
-      const targetInstance = wrapper.getInstanceByContextId(
-        contextId,
-        inquirerId,
-      );
-
-      targetInstance.instance = wrapper.forwardRef
-        ? Object.assign(targetInstance.instance, new metatype(...instances))
-        : new metatype(...instances);
+      instanceHost.instance = wrapper.forwardRef
+        ? Object.assign(
+            instanceHost.instance,
+            new (metatype as Type<any>)(...instances),
+          )
+        : new (metatype as Type<any>)(...instances);
     } else if (isInContext) {
       const factoryReturnValue = ((targetMetatype.metatype as any) as Function)(
         ...instances,
