@@ -92,7 +92,8 @@ export class ClientRMQ extends ClientProxy {
   }
 
   public createClient<T = any>(): T {
-    return rqmPackage.connect(this.urls) as T;
+    const socketOptions = this.getOptionsProp<RmqOptions>(this.options, 'socketOptions');
+    return rqmPackage.connect(this.urls, socketOptions) as T;
   }
 
   public mergeDisconnectEvent<T = any>(
