@@ -1,8 +1,21 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { join } from 'path';
 import { PhotoModule } from './photo/photo.module';
 
 @Module({
-  imports: [TypeOrmModule.forRoot(), PhotoModule],
+  imports: [
+    TypeOrmModule.forRoot({
+      type: 'mysql',
+      host: 'localhost',
+      port: 3306,
+      username: 'root',
+      password: 'root',
+      database: 'test',
+      entities: [join(__dirname, '**/**.entity{.ts,.js}')],
+      synchronize: true,
+    }),
+    PhotoModule,
+  ],
 })
 export class ApplicationModule {}
